@@ -1,13 +1,12 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net451;net461;netcoreapp2.0;</TargetFrameworks>
-    <AssemblyName>Serilog.Sinks.EventLog.Tests</AssemblyName>
+    <TargetFrameworks>net48;net6.0;net8.0</TargetFrameworks>
     <AssemblyOriginatorKeyFile>../../assets/Serilog.snk</AssemblyOriginatorKeyFile>
     <SignAssembly>true</SignAssembly>
     <PublicSign Condition=" '$(OS)' != 'Windows_NT' ">true</PublicSign>
-    <PackageId>Serilog.Sinks.EventLog.Tests</PackageId>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
+    <NoWarn>CA1416</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
@@ -20,23 +19,16 @@
     <ProjectReference Include="..\..\src\Serilog.Sinks.EventLog\Serilog.Sinks.EventLog.csproj" />
   </ItemGroup>
 
-  <ItemGroup>    
-      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
-      <PackageReference Include="NUnit3TestAdapter" Version="3.9.0" />
-      <PackageReference Include="Serilog" Version="2.3.0" />
-      <PackageReference Include="Serilog.Settings.Configuration" Version="2.3.1" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Xml" Version="1.0.0" />
-      <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
-      <PackageReference Include="NUnit" Version="3.9.0" />      
-  </ItemGroup>
-
   <ItemGroup>
-    <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+      <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
+      <PackageReference Include="xunit.runner.visualstudio" Version="2.8.1">
+          <PrivateAssets>all</PrivateAssets>
+          <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
+      </PackageReference>
+      <PackageReference Include="xunit" Version="2.8.1" />
+      <PackageReference Include="Serilog.Settings.Configuration" Version="8.0.1" />
+      <PackageReference Include="Microsoft.Extensions.Configuration.Xml" Version="8.0.0" />
+      <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net451' ">
-    <Reference Include="System" />
-    <Reference Include="Microsoft.CSharp" />
-  </ItemGroup>
-  
 </Project>
